fix(llm-dashboard): pin TZ in DST test + polish (delta/rank-key/dead-css)
